implement metal rendering pipeline

+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/View.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2010, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -751,12 +751,12 @@ public abstract class View {
         _end(this.ptr);
     }
 
-    protected abstract int _getNativeFrameBuffer(long ptr);
+    protected abstract long _getNativeFrameBuffer(long ptr);
 
     /**
      * Called on the renderer thread and must be between lock and unlock
      */
-    public int getNativeFrameBuffer() {
+    public long getNativeFrameBuffer() {
         return _getNativeFrameBuffer(this.ptr);
     }
