+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/gtk/GtkView.java
@@ -45,7 +45,7 @@ final class GtkView extends View {
     }
 
     @Override
-    protected int _getNativeFrameBuffer(long ptr) {
+    protected long _getNativeFrameBuffer(long ptr) {
         return 0;
     }
