+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/mac/MacView.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2010, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -66,7 +66,7 @@ final class MacView extends View {
         return multiClickMaxY;
     }
 
-    @Override native protected int _getNativeFrameBuffer(long ptr);
+    @Override native protected long _getNativeFrameBuffer(long ptr);
     @Override native protected long _create(Map caps);
     @Override native protected int _getX(long ptr);
     @Override native protected int _getY(long ptr);
