+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/win/WinView.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2010, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -66,7 +66,7 @@ final class WinView extends View {
     }
 
     @Override
-    protected int _getNativeFrameBuffer(long ptr) {
+    protected long _getNativeFrameBuffer(long ptr) {
         return 0;
     }
